fixes all room problems

+++ app/src/main/java/com/openclassrooms/realestatemanager/ui/map/MapFragment.kt
@@ -88,7 +88,7 @@ class MapFragment : BaseToolbarFragment(), MapView,
 
     override fun setMarkerClickListener() {
         googleMap.setOnInfoWindowClickListener { marker ->
-            presenter.onMarkerClicked(marker.tag.toString())
+            presenter.onMarkerClicked(marker.tag.toString().toInt())
         }
     }
